requirements.txt
py-clob-client==0.34.6
websockets>=12.0,<14.0
numpy>=1.26,<3.0
python-dotenv>=1.0,<2.0
aiohttp>=3.9,<4.0
